requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
google-adk>=0.1.0
feedparser>=6.0.10
litellm>=1.40.0

